try to make playback via MediaSession API work; another refactor
